Move ChannelType enum to avoid cyclic imports
