+++ src/client/channel_type.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class ChannelType(Enum):
+    LOG = "Action Log"
+    HOF = "Hall of Fame"
+    MODMAIL = "Anonymous Modmail"
+    SCHEDULES = "Schedules"
+    SUGGESTIONS = "Suggestions"
+    TWITCH = "Twitch"
